requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask_SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
mysqlclient==2.2.4
python-dotenv==1.0.1
nh3==0.2.18
Markdown==3.6